+++ src/pages/FaqNotice.tsx
@@ -1,6 +1,6 @@
 import Header from "@/components/layout/Header";
 import Footer from "@/components/layout/Footer";
-import FaqSection from "@/components/FaqSection";
+import FaqSection from "@/components/faq/Section";
 
 export default function FaqPage() {
   return (
